Add num_[reads|writes] to the jag_prec_t.
Not sure if this could be done a different way.  At the moment I
wouldn't mind making a filesystem specific structure to pass in. Or
perhaps make use of a different array or such.

In the end this might be the best way to do it, but it isn't clear at
the moment since the num_* isn't sent to the database in any fashion.

Bug 2782